Move classes in head_refinement.h/cpp to separate header/implementation files.
