+++ python/boomer/common/cpp/head_refinement/head_refinement_single.cpp
@@ -1,0 +1,89 @@
+#include "head_refinement_single.h"
+
+
+/**
+ * Allows to find the best single-label head that predicts for a single label.
+ *
+ * @tparam T The type of the vector that provides access to the indices of the labels that are considered when searching
+ *           for the best head
+ */
+template<class T>
+class SingleLabelHeadRefinement : public IHeadRefinement {
+
+    private:
+
+        const T& labelIndices_;
+
+        std::unique_ptr<PartialPrediction> headPtr_;
+
+    public:
+
+        /**
+         * @param labelIndices A reference to an object of template type `T` that provides access to the indices of the
+         *                     labels that should be considered when searching for the best head
+         */
+        SingleLabelHeadRefinement(const T& labelIndices)
+            : labelIndices_(labelIndices) {
+
+        }
+
+        const AbstractEvaluatedPrediction* findHead(const AbstractEvaluatedPrediction* bestHead,
+                                                    IStatisticsSubset& statisticsSubset, bool uncovered,
+                                                    bool accumulated) override {
+            const LabelWiseEvaluatedPrediction& prediction = statisticsSubset.calculateLabelWisePrediction(uncovered,
+                                                                                                           accumulated);
+            uint32 numPredictions = prediction.getNumElements();
+            LabelWiseEvaluatedPrediction::quality_score_const_iterator qualityScoreIterator =
+                prediction.quality_scores_cbegin();
+            uint32 bestC = 0;
+            float64 bestQualityScore = qualityScoreIterator[bestC];
+
+            for (uint32 c = 1; c < numPredictions; c++) {
+                float64 qualityScore = qualityScoreIterator[c];
+
+                if (qualityScore < bestQualityScore) {
+                    bestQualityScore = qualityScore;
+                    bestC = c;
+                }
+            }
+
+            // The quality score must be better than that of `bestHead`...
+            if (bestHead == nullptr || bestQualityScore < bestHead->overallQualityScore) {
+                LabelWiseEvaluatedPrediction::score_const_iterator scoreIterator = prediction.scores_cbegin();
+                typename T::const_iterator indexIterator = labelIndices_.cbegin();
+
+                if (headPtr_.get() == nullptr) {
+                    headPtr_ = std::make_unique<PartialPrediction>(1);
+                }
+
+                PartialPrediction::score_iterator headScoreIterator = headPtr_->scores_begin();
+                PartialPrediction::index_iterator headIndexIterator = headPtr_->indices_begin();
+                headScoreIterator[0] = scoreIterator[bestC];
+                headIndexIterator[0] = indexIterator[bestC];
+                headPtr_->overallQualityScore = bestQualityScore;
+                return headPtr_.get();
+            }
+
+            return nullptr;
+        }
+
+        std::unique_ptr<AbstractEvaluatedPrediction> pollHead() override {
+            return std::move(headPtr_);
+        }
+
+        const EvaluatedPrediction& calculatePrediction(IStatisticsSubset& statisticsSubset, bool uncovered,
+                                                       bool accumulated) const override {
+            return statisticsSubset.calculateLabelWisePrediction(uncovered, accumulated);
+        }
+
+};
+
+std::unique_ptr<IHeadRefinement> SingleLabelHeadRefinementFactory::create(
+        const FullIndexVector& labelIndices) const {
+    return std::make_unique<SingleLabelHeadRefinement<FullIndexVector>>(labelIndices);
+}
+
+std::unique_ptr<IHeadRefinement> SingleLabelHeadRefinementFactory::create(
+        const PartialIndexVector& labelIndices) const {
+    return std::make_unique<SingleLabelHeadRefinement<PartialIndexVector>>(labelIndices);
+}
