+++ python/boomer/common/cpp/indices/index_vector_full.cpp
@@ -1,5 +1,6 @@
 #include "index_vector_full.h"
-#include "../head_refinement.h"
+#include "../head_refinement/head_refinement.h"
+#include "../head_refinement/head_refinement_factory.h"
 #include "../statistics.h"
 #include "../thresholds.h"
