+++ python/boomer/common/cpp/predictions.cpp
@@ -1,4 +1,5 @@
-#include "head_refinement.h"
+#include "head_refinement/head_refinement.h"
+#include "head_refinement/head_refinement_factory.h"
 #include "predictions.h"
 #include "statistics.h"
 #include "thresholds.h"
